+++ _/components/administration/php-version/actions/addSecondHand.action.php
@@ -131,7 +131,7 @@ if ($validator->validate ( $_POST )) {
 			saveUploadPhoto ( 'sh_image_5', $ses_id, $secondHand->id );
 		}
 		// remove session folder
-		if (file_exists ( "../images/tempory/" . $ses_id )) {
+		if ($ses_id != "" && file_exists ( "../images/tempory/" . $ses_id )) {
 			rmdir ( "../images/tempory/" . $ses_id );
 		}
 	}
@@ -193,7 +193,7 @@ function saveUploadPhoto($uploadId, $ses_id, $id) {
 		mkdir ( "../images/secondHand/" . $id . "/sm/" );
 	}
 	$img = new SimpleImage();
-	$img->load("../images/secondHand/" . $id . "/" . $_SESSION [$uploadId])->best_fit(400, 400)->save("../images/secondHand/" . $id . "/sm/" . $_SESSION [$uploadId]);
+	$img->load("../images/secondHand/" . $id . "/" . $_SESSION [$uploadId])->best_fit(600, 600)->save("../images/secondHand/" . $id . "/sm/" . $_SESSION [$uploadId]);
 }
 function checkUploadPhoto($uploadId, $ses_id) {
 // 	echo '$uploadId(begin): '.$uploadId.'<br>';
